support response headers and allow others response type

+++ buildgo/buildgo.go
@@ -21,15 +21,21 @@ func Generate(doc *openapi3.T) {
 }
 
 func WriteInfra() {
+	overridefile := helper.Proj.OverrideHandle
 
 	//prepare main
-	var mainbytes bytes.Buffer
-	mainfilepath := "./templates/go/main.gotxt"
-	mainsrc := helper.ReadFile(mainfilepath)
-	maintemplate := template.New("main")
-	maintemplate, _ = maintemplate.Parse(mainsrc)
-	_ = maintemplate.Execute(&mainbytes, helper.Proj)
-	helper.WriteFile("", "main.go", mainbytes.String())
+	//if file exists, but no override flag
+	if helper.CheckFileExists("", "main.go") && overridefile == false {
+		log.Warn("main.go exists, skip")
+	} else {
+		var mainbytes bytes.Buffer
+		mainfilepath := "./templates/go/main.gotxt"
+		mainsrc := helper.ReadFile(mainfilepath)
+		maintemplate := template.New("main")
+		maintemplate, _ = maintemplate.Parse(mainsrc)
+		_ = maintemplate.Execute(&mainbytes, helper.Proj)
+		helper.WriteFile("", "main.go", mainbytes.String())
+	}
 
 	//prepare go.mod
 	var gomodbytes bytes.Buffer
